Client Server
First edition, client server project
we recive data from client to server, make query to sql  and get back the nearest point in the map.
The point sent back to the client.
All data decoded with XOR cryptograph and only decoded data can be transfered.

Client have a login frame, where he can sing up & sing in.

In the next version the sing in and sing up will be with the server, and we will get prefernens for client and for the server.

The project write with JAVA, SQL.
In the project i practice JAV, OOP, multithreading, SQL API, GUI, datastructure, cryptograph, and more else other tools.

The roject written by computer scince studetn in first half way to computer scince graduation.